Interative bouding box selecting for show_tracker_streamer
The first bouding box of object to track in show_tracker_streamer
is fixed. In this commit, we added an interative process that user
can select the first bouding box for tracking.

A problem occurs in this commit. After selecting the first bouding
box, the tracker will delay for a while. We should solve the
problem in future commits.