Modify show_regressor_test
Originally, show_regressor_test takes two crooped images and feeds
them to Regressor to estimate. Some problems occured when using
such method:

1. The cropped images may not follow the cropping rules of GOTURN.
2. Manually crooping images is a time consuming task.

In this commit, we solved the problems by modifying the process
of show_regressor_test. It takes two entire images and ground
truth of locations of the tracked object, and then it crops the
images automatically by following the rules of GOTURN and feeds
the cropped Regressor to estimate.